filter pagination

+++ app/src/main/java/com/badcompany/pitakpass/domain/model/DriverOffer.kt
@@ -6,4 +6,4 @@ data class PassengerOffer(@SerializedName("postId") val postId: Long,
                           @SerializedName("price") val price: Int? = null,
                           @SerializedName("message") val message: String? = null,
                           @SerializedName("seat") val seat: Int? = null,
-                          @SerializedName("repliedPostId") val repliedPostId: Int? = null)
+                          @SerializedName("repliedPostId") val repliedPostId: Long? = null)
